requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
jsonschema==4.23.0
firebase-admin>=6.0.0
